% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerSignComp-package.R
\docType{package}
\name{CancerSignComp-package}
\alias{CancerSignComp}
\alias{CancerSignComp-package}
\title{CancerSignComp: Benchmarking Cancer Prognostic Gene Signatures Against Each Other and Clinical Staging}
\description{
Provides a standardized framework for evaluating and comparing cancer prognostic gene signatures. Computes risk scores using weighted mean, GSVA, or ssGSEA methods; benchmarks multiple signatures against each other via concordance index (C-index); and compares signatures to TNM/AJCC clinical staging using Net Reclassification Improvement (NRI) and Integrated Discrimination Improvement (IDI). Includes Kaplan-Meier, C-index bar plot, and subgroup forest plot visualizations, as well as CSV and Excel export utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bernahukkamli/CancerSignComp}
  \item Report bugs at \url{https://github.com/bernahukkamli/CancerSignComp/issues}
}

}
\author{
\strong{Maintainer}: Berna Hukkamli \email{berna.hukkamli@sinop.edu.tr}

}
\keyword{internal}
