% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_signature.R
\name{print.CancerSignature}
\alias{print.CancerSignature}
\title{Print method for CancerSignature objects}
\usage{
\method{print}{CancerSignature}(x, ...)
}
\arguments{
\item{x}{A \code{CancerSignature} object}

\item{...}{Further arguments (ignored)}
}
\description{
Print method for CancerSignature objects
}
