% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_external.R
\name{print.ExternalValidation}
\alias{print.ExternalValidation}
\title{Print method for ExternalValidation objects}
\usage{
\method{print}{ExternalValidation}(x, ...)
}
\arguments{
\item{x}{An ExternalValidation object}

\item{...}{Further arguments (ignored)}
}
\description{
Print method for ExternalValidation objects
}
