% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_signature.R
\name{print.ScoredSignature}
\alias{print.ScoredSignature}
\title{Print method for ScoredSignature objects}
\usage{
\method{print}{ScoredSignature}(x, ...)
}
\arguments{
\item{x}{A \code{ScoredSignature} object}

\item{...}{Further arguments (ignored)}
}
\description{
Print method for ScoredSignature objects
}
