% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_signatures.R
\name{print.SignatureComparison}
\alias{print.SignatureComparison}
\title{Print method for SignatureComparison objects}
\usage{
\method{print}{SignatureComparison}(x, ...)
}
\arguments{
\item{x}{A \code{SignatureComparison} object}

\item{...}{Further arguments (ignored)}
}
\description{
Print method for SignatureComparison objects
}
