% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_staging.R
\name{print.StagingComparison}
\alias{print.StagingComparison}
\title{Print method for StagingComparison objects}
\usage{
\method{print}{StagingComparison}(x, ...)
}
\arguments{
\item{x}{A \code{StagingComparison} object}

\item{...}{Further arguments (ignored)}
}
\description{
Print method for StagingComparison objects
}
